+++ packages/perseus/src/widget-ai-utils/orderer/orderer-ai-utils.test.ts
@@ -1,5 +1,7 @@
 import {renderQuestion} from "../../widgets/__testutils__/renderQuestion";
 
+import {getPromptJSON} from "./orderer-ai-utils";
+
 import type {PerseusRenderer} from "../../perseus-types";
 
 const question1: PerseusRenderer = {
@@ -30,7 +32,45 @@ const question1: PerseusRenderer = {
     },
 };
 
-describe("orderer widget", () => {
+describe("Orderer AI utils", () => {
+    it("it returns JSON with the expected format and fields", () => {
+        const renderProps: any = {
+            options: [
+                {
+                    content: "Third item",
+                    images: {},
+                    widgets: {},
+                },
+                {
+                    content: "First item",
+                    images: {},
+                    widgets: {},
+                },
+                {
+                    content: "Second item",
+                    images: {},
+                    widgets: {},
+                },
+            ],
+        };
+
+        const userInput = {
+            current: ["First item", "Second item", "Third item"],
+        };
+
+        const resultJSON = getPromptJSON(renderProps, userInput);
+
+        expect(resultJSON).toEqual({
+            type: "orderer",
+            options: {
+                options: ["Third item", "First item", "Second item"],
+            },
+            userInput: {
+                values: ["First item", "Second item", "Third item"],
+            },
+        });
+    });
+
     it("should get prompt json which matches the state of the UI", async () => {
         // Arrange
         const {renderer} = renderQuestion(question1);
