+++ packages/perseus/src/widgets/orderer/orderer.tsx
@@ -12,7 +12,7 @@ import {Log} from "../../logging/log";
 import {ClassNames as ApiClassNames} from "../../perseus-api";
 import Renderer from "../../renderer";
 import Util from "../../util";
-import {getPromptJSON as _getPromptJSON} from "../../widget-ai-utils/orderer/prompt-utils";
+import {getPromptJSON as _getPromptJSON} from "../../widget-ai-utils/orderer/orderer-ai-utils";
 
 import {scoreOrderer} from "./score-orderer";
 
@@ -27,7 +27,7 @@ import type {
     PerseusOrdererRubric,
     PerseusOrdererUserInput,
 } from "../../validation.types";
-import type {OrdererPromptJSON} from "../../widget-ai-utils/orderer/prompt-utils";
+import type {OrdererPromptJSON} from "../../widget-ai-utils/orderer/orderer-ai-utils";
 import type {LinterContextProps} from "@khanacademy/perseus-linter";
 
 type PlaceholderCardProps = {
